# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Keycloak Middleware",
        "description": "Documentación Swagger de la API intermedia entre Laravel y Keycloak",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Realm Roles"
                ],
                "summary": "Listar roles del reino",
                "operationId": "d65ad88054aa19a95849109a46bfab0a",
                "parameters": [
                    {
                        "name": "realm",
                        "in": "query",
                        "description": "Nombre del reino",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/role-by-name": {
            "get": {
                "tags": [
                    "Realm Roles"
                ],
                "summary": "Obtiene rol por nombre del reino.",
                "operationId": "63a8bcc95d6f731c5b14dacf9adb6cac",
                "parameters": [
                    {
                        "name": "realm",
                        "in": "query",
                        "description": "Id del reino",
                        "required": true
                    },
                    {
                        "name": "role_name",
                        "in": "query",
                        "description": "Id del client",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/realms/realm-by-name": {
            "get": {
                "tags": [
                    "Realm"
                ],
                "summary": "Obtiene reino por nombre",
                "operationId": "3d8114084358ba290db326123341897c",
                "parameters": [
                    {
                        "name": "realm_name",
                        "in": "query",
                        "description": "Id del client",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer Token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}